Update signal usage in sign package
